More logging to debug a Neo4j/Heroku error
The traversing of the graph in Neo4j seems to currently be failing at my RelSpider Heroku instance - same code as before, but for some reason the Neo4j database is negliecting it. Added these additional logging to better understand what the issue is all about and I have contacted the Neo4j team with this information and hopefully a solution will be found.