Defined module versions more narrowly
Previously only the tested version was defined - basically to just keep track of what works and what doesn't - but since eg. Heroku installs packages itself on eg. every single little deploy I found it necessary to define the versions a bit more narrowly so that only updates resulting in non-breaking changes can be expected. This makes it a bit harder for me to discover new versions of modules(?) but should ensure better stability.