Now shutdown gracefully on eg. Heroku
Now initiates a shutdown when noticing a SIGTERM - it shuts down the HTTP server of the web workers and stops any new URL:s from being crawled and once the last URL has been crawled connections to database and cache systems are closed. In most cases this will hopefully happen within in 10 seconds and thus result in a graceful shutdown on Heroku servers so that restarts of scripts there works as smoothly as possible.